+++ nb_size.c
@@ -6,7 +6,7 @@
 /*   By: tholzheu <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/01 15:47:41 by tholzheu          #+#    #+#             */
-/*   Updated: 2018/11/01 16:45:25 by tholzheu         ###   ########.fr       */
+/*   Updated: 2018/11/01 19:41:26 by tholzheu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ size_t		signed_nb_size(long long nb)
 	return (count);
 }
 
-size_t		unsigned_nb_size(unsigned long long nb, int base)
+size_t		unsigned_nb_size(unsigned long long nb, unsigned long long base)
 {
 	size_t		count;
